requirements.txt
python-dotenv==1.0.0
requests==2.31.0
polars==0.20.31
